+++ src/RxInfer.jl
@@ -25,6 +25,7 @@ include("constraints/form/form_fixed_marginal.jl")
 include("constraints/form/form_point_mass.jl")
 include("constraints/form/form_sample_list.jl")
 
+include("callbacks/events.jl")
 include("callbacks/benchmark.jl")
 include("callbacks/trace.jl")
 include("callbacks/stop_early.jl")